+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # <img src="https://uploads-ssl.webflow.com/5ea5d3315186cf5ec60c3ee4/5edf1c94ce4c859f2b188094_logo.svg" alt="Pip.Services Logo" width="200"> <br/> MongoDB components for Python Changelog
 
+## <a name="3.1.0"></a> 3.1.0 (2021-05-26)
+
+### Bug fixes
+* fixed method names
+* added docstrings and examples
+
+### Features
+* Added type hints
+
+
+## <a name="3.0.1"></a> 3.0.1 (2021-04-14)
+* Update references
+
 ## <a name="3.0.0"></a> 3.0.0 (2020-12-24)
 
 Initial public release